requirements.txt
pandas
numpy
scikit-learn
pyyaml
xgboost
matplotlib
seaborn
uvicorn
fastapi
jinja2
python-multipart